+++ pmd-xml/src/main/java/net/sourceforge/pmd/lang/xml/rule/DomXPathRule.java
@@ -128,8 +128,6 @@ public class DomXPathRule extends AbstractRule {
     public DomXPathRule() {
         definePropertyDescriptor(XPATH_EXPR);
         definePropertyDescriptor(DEFAULT_NS_URI);
-        // for compatibility, but is ignored.
-        definePropertyDescriptor(XPathRule.VERSION_DESCRIPTOR);
     }